+++ src/app/mission/[id]/stopwatch/ButtonSection.tsx
@@ -8,7 +8,6 @@ import {
   useStopwatchTimeContext,
 } from '@/app/mission/[id]/stopwatch/Stopwatch.context';
 import Button from '@/components/Button/Button';
-import Stopwatch from '@/components/Stopwatch/Stopwatch';
 import { EVENT_LOG_CATEGORY, EVENT_LOG_NAME } from '@/constants/eventLog';
 import { eventLogger } from '@/utils';
 import {
@@ -20,19 +19,12 @@ import {
 } from '@/utils/storage/progressMission';
 import { css, cx } from '@styled-system/css';
 
-interface Props {
-  missionId: string;
-  missionName: string;
-}
-
-function StopwatchSection({ missionId, missionName }: Props) {
+function ButtonSection({ missionId }: { missionId: string }) {
   const { step, onNextStep } = useStopwatchStepContext();
-  const { minutes, seconds, time } = useStopwatchTimeContext();
+  const { minutes, time } = useStopwatchTimeContext();
 
-  const { isPending: isStopwatchPending } = useInitTimeSetting({ missionId });
   const { onInitStart, onMidStart, onStop, onMidOut, onFinish } = useStopwatch(missionId);
-
-  const stepper = time < 60 ? 0 : Math.floor(time / 60 / 10);
+  const { isPending: isStopwatchPending } = useInitTimeSetting({ missionId });
 
   const onStart = () => {
     if (time > 0) {
@@ -40,97 +32,60 @@ function StopwatchSection({ missionId, missionName }: Props) {
       return;
     }
     onInitStart();
-    // onNextStep('progress');
-
-    // // 이전 미션 기록 삭제 - 강제 접근 이슈
-    // checkPrevProgressMission(missionId);
-    // setMissionTimeStack(missionId, 'start');
-
-    // // 중도 재시작
-    // if (time > 0) {
-    //   eventLogger.logEvent(EVENT_LOG_NAME.STOPWATCH.CLICK_RESTART, EVENT_LOG_CATEGORY.STOPWATCH);
-    //   return;
-    // }
-    // // 초기시작
-    // eventLogger.logEvent(EVENT_LOG_NAME.STOPWATCH.CLICK_START, EVENT_LOG_CATEGORY.STOPWATCH);
-    // setMissionData(missionId);
   };
 
   const onFinishButtonClick = () => {
-    // onNextStep('stop');
-
     // 10분 지나기 전 끝내기 눌렀을 때
     if (Number(minutes) < 10) {
       onMidOut();
       return;
-      // eventLogger.logEvent(
-      //   EVENT_LOG_NAME.STOPWATCH.CLICK_FINISH_BUTTON_BEFORE_10MM,
-      //   EVENT_LOG_CATEGORY.STOPWATCH,
-      //   logData,
-      // );
-      // openMidOutModal();
-      // return;
     }
 
     onFinish();
-    // eventLogger.logEvent(EVENT_LOG_NAME.STOPWATCH.CLICK_FINISH_BUTTON, EVENT_LOG_CATEGORY.STOPWATCH, logData);
-    // openFinalModal();
   };
 
   return (
-    <>
-      <section className={opacityAnimation}>
-        <Stopwatch
-          minutes={minutes}
-          seconds={seconds}
-          missionName={missionName}
-          stepper={stepper}
-          isProgress={step === 'progress'}
-          isDisabled={step === 'stop'}
-        />
-      </section>
-      <section className={cx(buttonContainerCss, opacityAnimation)}>
-        {step === 'ready' && (
-          <div className={fixedButtonContainerCss}>
-            <Button variant="primary" size="large" type="button" onClick={onStart} disabled={isStopwatchPending}>
-              시작
-            </Button>
-          </div>
-        )}
-        {step === 'progress' && (
-          <>
-            <Button size="medium" variant="secondary" type="button" onClick={onStop}>
-              일시 정지
-            </Button>
-            <Button size="medium" variant="primary" type="button" onClick={onFinishButtonClick}>
-              끝내기
-            </Button>
-          </>
-        )}
-        {step === 'stop' && (
-          <>
-            <Button
-              size="medium"
-              variant="secondary"
-              type="button"
-              onClick={() => {
-                setMissionTimeStack(missionId, 'restart');
-                onNextStep('progress');
-              }}
-            >
-              다시 시작
-            </Button>
-            <Button size="medium" variant="primary" type="button" onClick={onFinishButtonClick}>
-              끝내기
-            </Button>
-          </>
-        )}
-      </section>
-    </>
+    <section className={cx(buttonContainerCss, opacityAnimation)}>
+      {step === 'ready' && (
+        <div className={fixedButtonContainerCss}>
+          <Button variant="primary" size="large" type="button" onClick={onStart} disabled={isStopwatchPending}>
+            시작
+          </Button>
+        </div>
+      )}
+      {step === 'progress' && (
+        <>
+          <Button size="medium" variant="secondary" type="button" onClick={onStop}>
+            일시 정지
+          </Button>
+          <Button size="medium" variant="primary" type="button" onClick={onFinishButtonClick}>
+            끝내기
+          </Button>
+        </>
+      )}
+      {step === 'stop' && (
+        <>
+          <Button
+            size="medium"
+            variant="secondary"
+            type="button"
+            onClick={() => {
+              setMissionTimeStack(missionId, 'restart');
+              onNextStep('progress');
+            }}
+          >
+            다시 시작
+          </Button>
+          <Button size="medium" variant="primary" type="button" onClick={onFinishButtonClick}>
+            끝내기
+          </Button>
+        </>
+      )}
+    </section>
   );
 }
 
-export default StopwatchSection;
+export default ButtonSection;
 
 const useStopwatch = (missionId: string) => {
   const { onNextStep } = useStopwatchStepContext();
@@ -196,6 +151,30 @@ const useStopwatch = (missionId: string) => {
   };
 };
 
+const buttonContainerCss = css({
+  margin: '28px auto',
+  display: 'flex',
+  justifyContent: 'center',
+  gap: '12px',
+});
+
+const opacityAnimation = css({
+  animation: 'fadeIn .7s',
+});
+
+const fixedButtonContainerCss = css({
+  position: 'fixed',
+  left: '16px',
+  right: '16px',
+  bottom: '16px',
+  width: '100%',
+  maxWidth: 'calc(475px  - 48px)',
+  margin: '0 auto',
+  '@media (max-width: 475px)': {
+    maxWidth: 'calc(100vw  - 48px)',
+  },
+});
+
 const MAX_SECONDS = 3600; // max 1 hour
 
 const useInitTimeSetting = ({ missionId }: { missionId: string }) => {
@@ -235,27 +214,3 @@ const useInitTimeSetting = ({ missionId }: { missionId: string }) => {
 
   return { isPending };
 };
-
-const buttonContainerCss = css({
-  margin: '28px auto',
-  display: 'flex',
-  justifyContent: 'center',
-  gap: '12px',
-});
-
-const opacityAnimation = css({
-  animation: 'fadeIn .7s',
-});
-
-const fixedButtonContainerCss = css({
-  position: 'fixed',
-  left: '16px',
-  right: '16px',
-  bottom: '16px',
-  width: '100%',
-  maxWidth: 'calc(475px  - 48px)',
-  margin: '0 auto',
-  '@media (max-width: 475px)': {
-    maxWidth: 'calc(100vw  - 48px)',
-  },
-});
